refactor(agent-campaign): extract winner analysis
